moved staking to the lending protocol; added v2 Algofi vault
