+++ projects/algofi/index.js
@@ -1,6 +1,6 @@
 const sdk = require('@defillama/sdk')
 const { toUSDTBalances } = require('../helper/balances')
-const axios = require("axios");
+const { get } = require('../helper/http')
 const { lookupApplications, } = require("../helper/algorand");
 
 const marketStrings = {
@@ -8,6 +8,7 @@ const marketStrings = {
     underlying_borrowed : "ub",
     underlying_reserves : "ur",
     active_collateral : "acc",
+    active_collateral_v2 : "ac",
     lp_circulation: "lc",
     bank_to_underlying_exchange: "bt",
     b_asset_circulation: "bac",
@@ -16,6 +17,35 @@ const marketStrings = {
     oracle_app_id: "oai",
 }
 
+const marketV2Strings = {
+    b_asset_to_underlying_exchange_rate: "baer"
+}
+
+const stakingV2Strings = {
+    total_staked: "ts",
+}
+
+const fixedValueStakingContracts = ["TINYMAN11_STBL_USDC_LP_STAKING", "ALGOFI-STBL-USDC-LP"]
+const singleSideStakingContracts = ["DEFLY", "STBL", "OPUL"]
+const variableValueStakingContracts = [
+                                        "ALGOFI-STBL-ALGO-LP",
+                                        "AF-XET-STBL-75BP-STAKING",
+                                        "AF-GOBTC-STBL-25BP-STAKING",
+                                        "AF-GOETH-STBL-25BP-STAKING",
+                                        "AF-OPUL-STBL-75BP-STAKING",
+                                        "AF-DEFLY-STBL-75BP-STAKING",
+                                        "AF-NANO-USDC-STBL-5BP-STAKING",
+                                        "AF-NANO-USDT-STBL-5BP-STAKING",
+                                        "AF-NANO-USDT-USDC-5BP-STAKING",
+                                        "AF-USDC-STBL-NANO-SUPER-STAKING",
+                                        "AF-ZONE-STBL-75BP-STAKING",
+                                        "AF-TINY-STBL-75BP-STAKING",
+                                        "AF-GOMINT-STBL-25BP-STAKING"
+                                      ]
+
+const stakingContractsV1 = fixedValueStakingContracts.concat(variableValueStakingContracts).concat(singleSideStakingContracts)
+const bankStakingContractsV2 = ["AF-BANK-USDC-STANDARD", "AF-BANK-USDt-STANDARD"]
+
 const orderedAssets = [
     "ALGO",
     "STBL",
@@ -23,6 +53,7 @@ const orderedAssets = [
     "goBTC",
     "goETH",
     "vALGO",
+    "vALGO2",
     "USDT",
     "STBL2",
     "AF-NANO-BSTBL2-BUSDC",
@@ -77,6 +108,12 @@ const appDictionary = {
     "vALGO": {
         "decimals": 6,
         "appIds": [465814318],
+        "oracleAppId": 531724540,
+        "oracleFieldName": "latest_twap_price",
+    },
+    "vALGO2": {
+        "decimals": 6,
+        "appIds": [879935316],
         "oracleAppId": 531724540,
         "oracleFieldName": "latest_twap_price",
     },
@@ -211,7 +248,6 @@ const appDictionary = {
     }
 }
 
-
 async function getAppGlobalState(marketId) {
   let response = await lookupApplications(marketId);
   let results = {}
@@ -239,7 +275,14 @@ async function getPrices(marketDictionary, orderedAssets) {
 }
 
 function getMarketSupply(assetName, appGlobalState, prices, appDictionary) {
-    let underlyingCash = ((assetName === "STBL") || (assetName === "vALGO")) ? appGlobalState[marketStrings.active_collateral] : appGlobalState[marketStrings.underlying_cash]
+    let underlyingCash = 0;
+    if (assetName == "vALGO2") {
+        underlyingCash = appGlobalState[marketStrings.active_collateral_v2];
+    } else if ((assetName === "STBL") || (assetName === "vALGO")) {
+        underlyingCash = appGlobalState[marketStrings.active_collateral];
+    } else {
+        underlyingCash = appGlobalState[marketStrings.underlying_cash];
+    }
     let supplyUnderlying = underlyingCash - appGlobalState[marketStrings.underlying_reserves]
     supplyUnderlying /= Math.pow(10, appDictionary[assetName]["decimals"])
 
@@ -251,6 +294,71 @@ function getMarketBorrow(assetName, appGlobalState, prices) {
     borrowUnderlying /= Math.pow(10, appDictionary[assetName]["decimals"])
 
     return borrowUnderlying * prices[assetName]
+}
+
+async function stakingV1() {
+    let lpCirculations = {}
+
+    let prices = {
+             "TINYMAN11_STBL_USDC_LP_STAKING": 2,
+             "ALGOFI-STBL-USDC-LP": 2,
+    }
+
+    for (const contractName of variableValueStakingContracts) {
+        let contractState = await getAppGlobalState(
+            appDictionary["STAKING_CONTRACTS"][contractName]["poolAppId"]
+        )
+        lpCirculations[contractName] = contractState[marketStrings.lp_circulation] / 1000000
+    }
+
+    let poolSnapshots = await get("https://thf1cmidt1.execute-api.us-east-2.amazonaws.com/Prod/amm_pool_snapshots/?network=MAINNET")
+    let assetSnapshots = await get("https://thf1cmidt1.execute-api.us-east-2.amazonaws.com/Prod/amm_asset_snapshots/?network=MAINNET")
+
+    for (const poolSnapshot of poolSnapshots.pool_snapshots) {
+        for (const contractName of variableValueStakingContracts) {
+            if (poolSnapshot.id == appDictionary["STAKING_CONTRACTS"][contractName]["poolAppId"]) {
+                prices[contractName] = poolSnapshot.balance_info.total_usd / lpCirculations[contractName]
+            }
+        }
+    }
+
+    for (const assetSnapshot of assetSnapshots.asset_snapshots) {
+        for (const contractName of singleSideStakingContracts) {
+            if (assetSnapshot.id == appDictionary["STAKING_CONTRACTS"][contractName]["assetId"]) {
+                prices[contractName] = assetSnapshot.price
+            }
+        }
+    }
+
+    let staked = 0
+    for (const contractName of stakingContractsV1) {
+        let appGlobalState = await getAppGlobalState(appDictionary["STAKING_CONTRACTS"][contractName]["appId"])
+        staked += getMarketSupply(contractName, appGlobalState, prices, appDictionary["STAKING_CONTRACTS"])
+    }
+
+    return toUSDTBalances(staked)
+}
+
+async function stakingV2() {
+    let totalStaked = 0
+    for (const contractName of bankStakingContractsV2) {
+        let stakingAppGlobalState = await getAppGlobalState(appDictionary["STAKING_CONTRACTS"][contractName]["appId"])
+        let marketAppGlobalState = await getAppGlobalState(appDictionary["STAKING_CONTRACTS"][contractName]["marketAppId"])
+
+        let bAssetStaked = stakingAppGlobalState[stakingV2Strings.total_staked] / 1000000
+        let bAssetToUnderlyingExchange = marketAppGlobalState[marketV2Strings.b_asset_to_underlying_exchange_rate] / 1000000000
+        let underlyingStaked = bAssetStaked * bAssetToUnderlyingExchange
+
+        let oracleState = await getAppGlobalState(marketAppGlobalState[marketStrings.oracle_app_id])
+
+        let oraclePriceFieldName = appDictionary["STAKING_CONTRACTS"][contractName]["oracleFieldName"]
+        let oraclePrice = oracleState[oraclePriceFieldName] / 1000000
+
+        let stakedUsd = underlyingStaked * oraclePrice
+
+        totalStaked  += stakedUsd
+    }
+    return toUSDTBalances(totalStaked)
 }
 
 async function borrowed() {
@@ -277,6 +385,7 @@ async function supply() {
             let appGlobalState = await getAppGlobalState(id)
             let assetTvl = getMarketSupply(assetName, appGlobalState, prices, appDictionary)
             supply += assetTvl
+            console.log(assetName, assetTvl);
         }
     }
 
@@ -286,7 +395,8 @@ async function supply() {
 module.exports = {
     algorand: {
         tvl: supply,
-        borrowed
+        borrowed,
+        staking: sdk.util.sumChainTvls([stakingV1, stakingV2])
     }
 }
 // node test.js projects/algofi/index.js
